requirements.txt
pytorch_lightning==1.0.8
hydra-core==1.0.4
sweetviz==2.0.2
comet_ml==3.2.7
jpholiday
pandas==1.1.4
scipy==1.5.4
matplotlib==3.3.3
seaborn==0.11.0
scikit-learn==0.23.2
opencv-python==4.4.0.46
category_encoders==2.2.2
catboost==0.24.3
umap-learn==0.4.6
mojimoji==0.0.11